+++ packages/terra-docs/tests/wdio/button-spec.js
@@ -2,7 +2,7 @@ Terra.describeViewports('Button', ['medium'], () => {
   describe('Default Button', () => {
     it('is valid', () => {
       browser.url('/raw/tests/terra-docs/docs/button');
-      browser.moveToObject('#root', 700, 700);
+      $('#root').moveTo({ xOffset: 700, yOffset: 700 });
       Terra.validates.element('is valid-default', { selector: '#site' });
     });
 
@@ -18,7 +18,7 @@ Terra.describeViewports('Button', ['medium'], () => {
 
     it('checks mouse interactions', () => {
       browser.url('/raw/tests/terra-docs/docs/button');
-      browser.moveToObject('#button');
+      $('#button').moveTo();
       Terra.validates.element('checks mouse interactions - hover', { selector: '#site' });
       browser.buttonDown();
       Terra.validates.element('checks mouse interactions - active', { selector: '#site' });