+++ packages/terra-dialog/tests/wdio/dialog-spec.js
@@ -2,6 +2,6 @@ Terra.describeViewports('Dialog', ['tiny', 'large'], () => {
   it('displays default Dialog', () => {
     browser.url('/raw/tests/terra-dialog/dialog/default-dialog');
 
-    Terra.validates.element('default', { selector: '#dialog' });
+    Terra.validates.element('default dialog', { selector: '#dialog' });
   });
 });