+++ packages/terra-docs/tests/wdio/placeholder-spec.js
@@ -2,7 +2,7 @@ Terra.describeViewports('Placeholder', ['medium'], () => {
   describe('Default Placeholder', () => {
     it('checks accessibility', () => {
       browser.url('/raw/tests/terra-docs/docs/placeholder');
-      Terra.validates.element();
+      Terra.validates.element('default');
     });
   });
 });