+++ packages/terra-form-select/src/multiple/Menu.jsx
@@ -206,6 +206,7 @@ class Menu extends React.Component {
 
   componentDidUpdate() {
     this.updateNoResultsScreenReader();
+    this.updateCurrentActiveScreenReader();
   }
 
   componentWillUnmount() {
@@ -375,6 +376,8 @@ class Menu extends React.Component {
       intl,
       visuallyHiddenComponent,
     } = this.props;
+    const options = MenuUtil.flatten(this.state.children, true);
+    const totalOptions = options.length;
 
     const clearSelectTxt = intl.formatMessage({ id: 'Terra.form.select.clearSelect' });
 
@@ -398,14 +401,15 @@ class Menu extends React.Component {
     }
 
     if (element) {
+      const index = MenuUtil.getIndex(options, this.state.active);
       if (element.props.display === '' && element.props.value === '') {
         // Used for case where users selects clear option and opens
         // dropdown again and navigates to clear option
         visuallyHiddenComponent.current.innerText = clearSelectTxt;
       } else if (this.isActiveSelected()) {
-        visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.selectedText' }, { text: displayText });
+        visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.selectedText' }, { text: displayText, index, totalOptions });
       } else {
-        visuallyHiddenComponent.current.innerText = displayText;
+        visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.activeOption' }, { text: displayText, index, totalOptions });
       }
     }
   }