+++ packages/terra-section-header/tests/wdio/section-header-spec.js
@@ -8,37 +8,37 @@ Terra.describeViewports('SectionHeader', ['huge'], () => {
   it('should display long title Section Header', () => {
     browser.url('/raw/tests/terra-section-header/section-header/long-title-section-header');
 
-    Terra.validates.element('long title');
+    Terra.validates.element('default long title');
   });
 
   describe('Long Title Accordion', () => {
     before(() => {
       browser.url('/raw/tests/terra-section-header/section-header/long-title-accordion-section-header');
-      browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
     });
 
     it('should display long title accordion Section Header', () => {
-      Terra.validates.element('default');
+      Terra.validates.element('long title accordion default');
     });
 
     it('should display long title accordion Section Header with hover', () => {
-      browser.moveToObject('.accordionContent');
+      $('.accordionContent').moveTo();
 
-      Terra.validates.element('hover');
+      Terra.validates.element('long title accordion hover');
     });
 
     it('should display keyboard focused long title accordion Section Header', () => {
-      browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
       browser.keys('Tab');
 
-      Terra.validates.element('keyboard focus');
+      Terra.validates.element('long title accordion keyboard focus');
     });
 
     it('should display active long title accordion Section Header', () => {
-      browser.moveToObject('.accordionContent');
+      $('.accordionContent').moveTo();
       browser.buttonDown();
 
-      Terra.validates.element('active');
+      Terra.validates.element('long title accordion active');
 
       browser.buttonUp();
     });
@@ -46,21 +46,21 @@ Terra.describeViewports('SectionHeader', ['huge'], () => {
 
   it('should display closed Section Header', () => {
     browser.url('/raw/tests/terra-section-header/section-header/closed-section-header');
-    browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
 
     Terra.validates.element('closed');
   });
 
   it('should display open Section Header', () => {
     browser.url('/raw/tests/terra-section-header/section-header/open-section-header');
-    browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
 
     Terra.validates.element('open');
   });
 
   it('should display onClick Section Header', () => {
     browser.url('/raw/tests/terra-section-header/section-header/on-click-section-header');
-    browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+    $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
 
     Terra.validates.element('onClick');
   });
@@ -75,23 +75,23 @@ Terra.describeViewports('SectionHeader', ['huge'], () => {
     before(() => browser.url('/raw/tests/terra-section-header/section-header/transparent-on-click-section-header'));
 
     it('should display hovered interactable transparent Section Header', () => {
-      browser.moveToObject('[data-id="section-header"]');
+      $('[data-id="section-header"]').moveTo();
 
-      Terra.validates.element('hover');
+      Terra.validates.element('interactable transparent hover');
     });
 
     it('should display keyboard focused interactable transparent Section Header', () => {
-      browser.moveToObject('#root', 0, 700); // move cursor to remove focus styles
+      $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
       browser.keys('Tab');
 
-      Terra.validates.element('keyboard focus');
+      Terra.validates.element('interactable transparent keyboard focus');
     });
 
     it('should display active interactable transparent Section Header', () => {
-      browser.moveToObject('[data-id="section-header"]');
+      $('[data-id="section-header"]').moveTo();
       browser.buttonDown();
 
-      Terra.validates.element('active');
+      Terra.validates.element('interactable transparent active');
     });
   });
 });