+++ packages/terra-show-hide/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+
 ## 2.48.0 - (February 2, 2021)
 
 * Changed
@@ -11,7 +14,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 2.46.0 - (December 8, 2020)
 
 * Changed
@@ -43,7 +46,7 @@
 
 * Added
   * Added an interaction media query to apply `hover` styles only on non-touch devices and removed hover wdio tests.
-  
+
 ## 2.40.0 - (August 6, 2020)
 
 * Changed